mysofra/serializers.py
from rest_framework import serializers
from mysofra.models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ('id', 'name', 'price', 'image_path', 'description')

